Mixing nested groups and examples
At this point I had to change the implementation to make it easy on my
poor brain. I've been trying to avoid putting too much into memory to
cope with large test suites, but the hairy cases are kind of difficult
to manage without maintaining some sort of state.

So I've decided that, given that we're using ndjson and that means we'll
only have one top-level example group in memory it once, it's probably
OK to just build a tree and then dump that tree out as JSON once the
next top level group comes along. That makes it a hell of a lot easier
to reason about. I'm sacrificing efficiency for developer happiness.